requirements.txt
asgiref==3.5.2
attrs==21.4.0
Django==4.0.6
djangorestframework==3.13.1
gdal==3.4.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
sqlparse==0.4.2
tomli==2.0.1
